alignment-microservice/requirements.txt
biopython==1.74
Click==7.0
colorama==0.3.9
docutils==0.15.2
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
numpy==1.17.1
pyasn1==0.4.7
PyYAML==5.1.2
rsa==3.4.2
urllib3==1.25.3
Werkzeug==0.15.6
